0928 - 결제 구현

+++ hoopi/src/main/java/com/ms/hoopi/admin/user/model/dto/UserOrderDetailSelectDto.java
@@ -15,5 +15,5 @@ public class UserOrderDetailSelectDto {
     private OrderDetailId orderDetailId;
     private Long quantity;
     private Long orderAmount;
-    private BigDecimal totalPrice;
+    private Long totalPrice;
 }
